feat(tt): orders transcript and saves ids
